requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
asyncpg==0.30.0